+++ src/main/config.js
@@ -26,7 +26,7 @@ export const EXTENSION_HASN = {
   pdf: '.pdf'
 }
 
-export const DEFAULT_THEME = 'dark'
+// export const DEFAULT_THEME = 'dark'
 
 export const VIEW_MENU_ITEM = {
   'Source Code Mode': false,