+++ src/main/menus/theme.js
@@ -1,18 +1,21 @@
 import * as actions from '../actions/theme'
+import { getUserPreference } from '../utils'
+
+const { theme } = getUserPreference()
 
 export default {
   label: 'Theme',
   submenu: [{
     label: 'Dark',
     type: 'radio',
-    checked: false,
+    checked: theme === 'dark',
     click (menuItem, browserWindow) {
       actions.selectTheme('dark')
     }
   }, {
     label: 'Light',
     type: 'radio',
-    checked: true,
+    checked: theme === 'light',
     click (menuItem, browserWindow) {
       actions.selectTheme('light')
     }